feat: library `canhttp` (#364)
Starts extracting the logic to handle HTTPs outcalls into a separate
library, named `canhttp`, that temporarily lives in the same repository.
The main idea is to start from a low-level client implementing the
[`tower::Service`](https://docs.rs/tower/latest/tower/trait.Service.html)
trait that can be extended by additional middlewares. This PR focusses
on this low-level client and a first middleware handling cycles
accounting.

Left for future PRs:
1. Add a mapping layer so that requests are of type `http::Request<>`
and responses are of type `http::Response<>`. This will allow to be able
to use the `Request` builder from the
[`http`](https://crates.io/crates/http) crate as well as various
HTTP-specific middlewares such as the ones from the
[`tower-http`](https://crates.io/crates/tower-http) crate. In
particular, the
[`TraceLayer`](https://docs.rs/tower-http/latest/tower_http/trace/index.html)
will allow to react on requests/responses to add the appropriates
metrics and logs.
2. Add an additional mapping layer for JSON-RPC applications so that
requests are of type `http::Request<JsonRpcRequest<>>` and responses are
of type `http::Response<JsonRpcResponse<>>` to easily make JSON-RPC
requests.

---------

Co-authored-by: Louis Pahlavi <louis.pahlavi@gmail.com>