feat(canhttp): JSON-RPC request ID with constant binary size (#397)
Add a wrapper around a JSON-RPC request ID to ensure that the amount of
bytes used when serialized to JSON is constant.

This is necessary to ensure that two HTTPs outcalls for two JSON-RPC
requests differing only by their IDs will have an identical cycle cost.
See also dfinity/sol-rpc-canister#62.