feat: Customize maximum block range for `eth_getLogs` (#424)
Allow callers to specify a larger block range to be queried via
`eth_getLogs`.

Most Ethereum JSON-RPC providers limits the block range that can be
queried for `eth_getLogs`. The concrete limit depends on the chosen
provider and until that PR the maximum block range was hardcoded to the
lowest limit (500 blocks) across all supported providers to guarantee
that the call will not be rejected.

Callers that explicitly specify which provider to query (inclusive
custom ones) may in particular wish to increase that limit if they know
that the new chosen value is supported by that provider.