feat!: add support for `eth_getTransactionReceipt` response fields (#474)
(XC-412) Add support for the `root` and `cumulativeGasUsed` fields in
the response to `eth_getTransactionReceipt`:
* The `root` field is mandatory for transactions included pre-Byzantium
and represents the post-transaction state root.
* The `cumulativeGasUsed` is mandatory for all transactions and
represents the sum of gas used by this transaction and all preceding
transactions in the same block.

This is **not** breaking for the canister Candid API since adding a new
field to a returned type is not breaking.

BREAKING: This is only breaking for the `evm_rpc_types` crate since it
adds new fields to the public `evm_rpc_types::TransactionsReceipt` type.